Revert "Update sherlock-platform to IntelliJ idea/251.26094.121 (#166)"
This reverts commit 854ce640269d6be52028226f600fe66fea5b2135.

+++ platform/editor-ui-api/src/com/intellij/openapi/actionSystem/toolbarLayout/AutoLayoutStrategy.kt
@@ -89,7 +89,7 @@ internal class AutoLayoutStrategy(private val myForceShowFirstComponent: Boolean
         val isLast = i == componentCount - 1
 
         val eachBound = Rectangle(getChildPreferredSize(component, i))
-        maxHeight = max(eachBound.height, maxHeight)
+        maxHeight = max(eachBound.height.toDouble(), maxHeight.toDouble()).toInt()
 
         if (!full) {
           val inside = if (isLast) eachX + eachBound.width <= widthToFit else eachX + eachBound.width + autoButtonSize.width <= widthToFit
